# docs/swagger.yaml
definitions:
  DatabaseServicev1.AddCardToCompanyRequest:
    properties:
      card:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.CreateCardCompanyRequest'
        description: '* Банковская карта'
    type: object
  DatabaseServicev1.AddCardToCompanyResponse:
    properties:
      address:
        description: '* Адрес офиса компании'
        type: string
      card:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.CardCompany'
        description: '* Банковская карта компании'
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID компании в базе данных'
        type: integer
      inn:
        description: '* ИНН юридического лица'
        type: string
      kpp:
        description: '* КПП юридического лица'
        type: string
      okpo:
        description: '* ОКПО предприятия/организации'
        type: string
      phone:
        description: '* Номер телефона компании'
        type: string
      site:
        description: '* Сайт компании'
        type: string
      title:
        description: '* Название компании'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя к которому относится данная компания'
        type: integer
    type: object
...
  DatabaseServicev1.CardCompany:
    properties:
      companyId:
        description: '* ID компании которой принадлежит данная карта'
        type: integer
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      cvv:
        description: '* CVV код карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      id:
        description: '* ID банковской карты в базе данных'
        type: integer
      number:
        description: '* Номер карты'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
    type: object
...
  DatabaseServicev1.CompaniesResponse:
    properties:
      companies:
        description: '* Массив компаний'
        items:
          $ref: '#/definitions/DatabaseServicev1.Company'
        type: array
    type: object
  DatabaseServicev1.Company:
    properties:
      address:
        description: '* Адрес офиса компании'
        type: string
      card:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.CardCompany'
        description: '* Банковская карта компании'
      createdAt:
        description: '* Дата создания сущности в базе данных'
        type: string
      id:
        description: '* ID компании в базе данных'
        type: integer
      inn:
        description: '* ИНН юридического лица'
        type: string
      kpp:
        description: '* КПП юридического лица'
        type: string
      okpo:
        description: '* ОКПО предприятия/организации'
        type: string
      phone:
        description: '* Номер телефона компании'
        type: string
      site:
        description: '* Сайт компании'
        type: string
      title:
        description: '* Название компании'
        type: string
      updatedAt:
        description: '* Дата последнего обновления сущности в базе данных'
        type: string
      userId:
        description: '* ID пользователя к которому относится данная компания'
        type: integer
    type: object
...
  DatabaseServicev1.CreateCardCompanyRequest:
    properties:
      companyId:
        description: '* ID компании которой принадлежит данная карта'
        type: integer
      cvv:
        description: '* CVV код банковской карты'
        type: integer
      date:
        description: '* Дата до которой активна карта'
        type: string
      fullName:
        description: '* ФИО с банковской карты'
        type: string
      number:
        description: '* Номер банковской карты'
        type: string
    type: object
...
  DatabaseServicev1.CreateCompanyRequest:
    properties:
      address:
        description: '* Адрес офиса компании'
        type: string
      card:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.CardCompany'
        description: '* Банковская карта компании'
      inn:
        description: '* ИНН юридического лица'
        type: string
      kpp:
        description: '* КПП юридического лица'
        type: string
      okpo:
        description: '* ОКПО предприятия/организации'
        type: string
      phone:
        description: '* Номер телефона компании'
        type: string
      site:
        description: '* Сайт компании'
        type: string
      title:
        description: '* Название компании'
        type: string
      userId:
        description: '* ID пользователя к которому относится данная компания'
        type: integer
    type: object
...
  DatabaseServicev1.DeleteCompanyByModelRequest:
    properties:
      company:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.Company'
        description: '* Полная сущность модели для удаления'
    type: object
...
  DatabaseServicev1.HTTPCodes:
    properties:
      code:
        description: '* HTTP код ответа'
        type: integer
    type: object
...
  DatabaseServicev1.UpdateCompanyRequest:
    properties:
      company:
        allOf:
        - $ref: '#/definitions/DatabaseServicev1.Company'
        description: '* Компания которую обновляем'
    type: object
...
  server.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервер маршрутизации
  title: API Gateway
  version: "1.0"
paths:
...
  /api/v1/companies:
    get:
      consumes:
      - application/json
      description: Массив компаний в базе данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CompaniesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Список всех компаний
      tags:
      - Company
    post:
      consumes:
      - application/json
      description: Создание новой сущности компании
      parameters:
      - description: Сущность компании
        in: body
        name: company
        schema:
          $ref: '#/definitions/DatabaseServicev1.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Создание компании
      tags:
      - Company
    put:
      consumes:
      - application/json
      description: Обновление сущности компании
      parameters:
      - description: Модель для обновления
        in: body
        name: company
        schema:
          $ref: '#/definitions/DatabaseServicev1.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Обновление компании
      tags:
      - Company
  /api/v1/companies/:
    get:
      consumes:
      - application/json
      description: Поиск компании по ее phone
      parameters:
      - description: Phone
        in: query
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Поиск компании по номеру телефона
      tags:
      - Company
  /api/v1/companies/{id}:
    get:
      consumes:
      - application/json
      description: Поиск компании по ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Поиск компании
      tags:
      - Company
  /api/v1/companies/{id}/card:
    get:
      consumes:
      - application/json
      description: Извлечение банковской карты компании по ее ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.CardCompany'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Извлечение банковской карты компании
      tags:
      - Company
  /api/v1/companies/addCard:
    post:
      consumes:
      - application/json
      description: Добавляет банковскую карту компании, поле companyId это ID компании
        в базе данных, которой будем добавлять карту
      parameters:
      - description: Сущность банковской карты
        in: body
        name: card
        schema:
          $ref: '#/definitions/DatabaseServicev1.AddCardToCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.AddCardToCompanyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Добавляет банковскую карту компании
      tags:
      - Company
  /api/v1/companies/deleteModel:
    post:
      consumes:
      - application/json
      description: Удаляет компании опираясь на всю сущность модели
      parameters:
      - description: Модель компании
        in: body
        name: company
        schema:
          $ref: '#/definitions/DatabaseServicev1.DeleteCompanyByModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление компании по модели
      tags:
      - Company
  /api/v1/company/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление компании по ID
      parameters:
      - description: ID компании
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DatabaseServicev1.HTTPCodes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      security:
      - BearerAuth: []
      summary: Удаление компании
      tags:
      - Company
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"